add working script usable as hotkey command
The are many options how to access clipboard from python script, but
using Tkinter module seems to be the easiest way for my system.

On X11 systems there is a known problem when clipboard content added by
script is lost as soon as the script ends. Add timeout as a workaround,
so after launching a script with hotkey you have 2 seconds to paste from
generated clipboard.